fix: styles in styled Components
