+++ app/globals.css
@@ -2,4 +2,5 @@ body {
   padding: 0px;
   margin: 0px;
   background-color: #f0f0f0;
+  height: 100%;
 }
