+++ pages/favorite.tsx
@@ -1,38 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { SearchBar } from "../components/SearchBar/SearchBar";
-import { styled } from "styled-components";
-import "../app/globals.css";
+import {
+  MainContainer,
+  ContainerCard,
+} from "../components/Favorite/styledFavorite";
 import { Card } from "@/components/Card/Card";
-
-const MainContainer = styled.div`
-  display: flex;
-  margin-top: 6rem;
-  margin-bottom: 6rem;
-  justify-content: center;
-
-  @media (max-width: 768px) {
-    flex-direction: column;
-    align-items: center;
-  }
-`;
-
-const ContainerCard = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-template-rows: 1fr;
-  gap: 30px 30px;
-  grid-auto-flow: row;
-  align-content: stretch;
-  justify-items: stretch;
-  width: 50%;
-  height: 50%;
-
-  @media (max-width: 768px) {
-    grid-template-columns: 1fr;
-    gap: 20px;
-    width: 80%;
-  }
-`;
+import "../app/globals.css";
 
 export default function Favorite() {
   const [favoriteSearch, setFavoriteSearch] = useState("");
